feat(client): ドロワー上部に額縁付き肖像画と称号を表示し業績画面へ遷移
- 業績画面の額縁付き肖像画を CavivaraPortrait として共有コンポーネント化
- ドロワーのヘッダーを肖像画＋称号キャプションに変更し、タップで
  業績画面へ遷移
- ドロワーでは内側の簡易的な額縁のみ描画し、発光とドロップシャドウは
  付けない（simplified オプション）
